Bound GFX_GetDisplayRefreshRate to maximum DOS rate (120 Hz)
There is no sense in going higher: it meets the highest minimum cut off
for VRR monitors (48Hz) and allows peak VESA rates of 120 Hz.

+++ src/ints/int10_modes.cpp
@@ -1231,7 +1231,7 @@ bool INT10_SetVideoMode(Bit16u mode)
 			misc_output|=0x0c;		//Select clock 3
 
 			// Use 70 Hz as the lower-bound
-			constexpr int vesa_min_refresh = 70;
+			constexpr int8_t vesa_min_refresh = 70;
 
 			// But allow a higher rate from the host and inform the user if so.
 			const auto refresh = std::max(vesa_min_refresh, GFX_GetDisplayRefreshRate());
